Add files via upload
This repo is a collection of Project 0 for the UC Berkely's Sp21 Data Strucutre class.
Thanks a lot for open source skeleton code on Github for me to finish the work.
(Skeleton code can be found on  repo:Berkeley-CS61B/skeleton-sp21 )

Run the Main program in game2048 in the project to see the result. 
Most of personal work/answers is at the Model part in game2048.